--- components/layout/Footer.tsx
@@ -1,9 +1,0 @@
-export default function Footer() {
-	return (
-		<footer className='footer footer-center h-16'>
-			<aside>
-				<p>Copyright © 2024 - All rights reserved by npSpeak</p>
-			</aside>
-		</footer>
-	);
-}
